Refactor: 게시글 상세, 게시글 내용 컴포넌트 개편
- articleId 대신 article을 props로 전달
- article을 가져오는 부분 react-query 적용
- 게시글 좋아요 컴포넌트 명시적으로 rename
- components 디렉토리로 이동
- 범용성을 위해 시간 util 함수 getCreatedAt으로 rename
