+++ src/components/pages/articles/ArticleDetailPage/components/ArticleContainer.tsx
@@ -2,23 +2,19 @@ import { useGetArticleById } from '@common/hooks/api/article';
 import { getCategoryName, useGetCategory } from '@common/hooks/api/category';
 import ArticleDetailHeader from './ArticleDetailHeader';
 import ArticleContent from './ArticleContent';
+import { Article } from '@root/network/types/Article';
 
 interface ArticleContainer {
-  articleId: number;
+  article: Article;
 }
 
-const ArticleContainer = ({ articleId }: ArticleContainer) => {
+const ArticleContainer = ({ article }: ArticleContainer) => {
   const { categories } = useGetCategory();
-  const { isError, article } = useGetArticleById(articleId);
   const categoryName = getCategoryName(categories, article?.categoryId);
   return (
     <>
-      {article && (
-        <>
-          <ArticleDetailHeader article={article} categoryName={categoryName} />
-          <ArticleContent article={article} />
-        </>
-      )}
+      <ArticleDetailHeader article={article} categoryName={categoryName} />
+      <ArticleContent article={article} />
     </>
   );
 };
