+++ src/components/pages/articles/ArticleDetailPage/components/ArticleContent.tsx
@@ -1,10 +1,10 @@
 import { ArticleProps } from '@components/pages/articles/common/types';
-import Like from './Like';
+import ArticleLike from './ArticleLike';
 
 const ArticleContent = ({ article }: ArticleProps) => (
   <div>
     <div>글내용 {article.content}</div>
-    <Like article={article} />
+    <ArticleLike article={article} />
   </div>
 );
