+++ src/components/pages/articles/ArticleDetailPage/components/ArticleDetailInfo.tsx
@@ -1,11 +1,11 @@
 import { ArticleProps } from '@components/pages/articles/common/types';
-import { getArticleTime } from '@components/pages/articles/common/utils';
+import { getCreatedAt } from '@components/pages/articles/common/utils';
 
 const ArticleDetailInfo = ({ article }: ArticleProps) => {
   return (
     <div>
       글정보
-      <div>⏱ {getArticleTime(article.createdAt)}</div>
+      <div>⏱ {getCreatedAt(article.createdAt)}</div>
       <div>👀 {article.viewCount}</div>
       <div>💬 {article.likeCount}</div>
     </div>
