+++ src/components/pages/articles/ArticleDetailPage/components/ArticleUpdateDelete.tsx
@@ -9,6 +9,7 @@ const ArticleUpdateDelete = ({ articleId, categoryId }: IdProps) => {
   const navigate = useNavigate();
   const { article } = useGetArticleById(articleId);
   const handleClickUpdate = () => {
+    // TODO: 수정 페이지로 이동
     navigate(`${URLs.WRITING}`, { state: article });
     // navigate(`${URLs.WRITING}/${articleId}`);
   };
