+++ src/components/pages/articles/ArticleDetailPage/index.jsx
@@ -1,32 +1,24 @@
-import { Body } from '@root/components/atoms/Board';
-import { ArticleService } from '@root/network';
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
+
+import { Body } from '@root/components/atoms/Board';
+import { ArticleService } from '@root/network';
 import Board from '../common/Board';
-import ArticleContainer from './ArticleContainer';
-import ArticleContent from './ArticleContentOld';
-import Comment from './Comment';
+import ArticleContainer from './components/ArticleContainer';
+import CommentContainer from './components/CommentContainer';
+import { useGetArticleById } from '@root/common/hooks/api/article';
 
 const ArticleDetailPage = () => {
-  const params = useParams();
-  const { id } = params;
-
-  const [article, setArticle] = useState();
-
-  useEffect(() => {
-    (async () => {
-      let data = await ArticleService.getArticleById(id);
-      setArticle(data);
-    })();
-  }, []);
+  const { id } = useParams();
+  const { article } = useGetArticleById(id);
 
   return (
     <Board categoryId={article?.categoryId}>
       <div className="block article_block">
         {article ? (
           <>
-            <ArticleContainer articleId={id} />
-            <Comment articleId={id} writer={article.writer} />
+            <ArticleContainer article={article} />
+            {/* <CommentContainer articleId={article.id} /> */}
           </>
         ) : (
           // TODO: 이쁘게 만들기
