+++ src/components/pages/articles/common/utils.tsx
@@ -1,8 +1,8 @@
 import dayjs from 'dayjs';
 
-type GetArticleTime = (time: Date) => string;
+type GetCreatedAt = (time: Date) => string;
 
-export const getArticleTime: GetArticleTime = time => {
+export const getCreatedAt: GetCreatedAt = time => {
   const dayjsTime = dayjs(time);
   return dayjsTime.isSame(dayjs(), 'day') ? dayjsTime.format('HH:mm') : dayjsTime.format('MM/DD');
 };
